pyproject.toml
[project]
name = "social-ai-studio"
version = "0.4.0"
description = "Social AI Studio — AI-Powered Social Media Content Studio for TechConnect 2026 Hackathon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agent-framework-core",
    "azure-identity",
    "azure-cosmos",
    "azure-monitor-opentelemetry",
    "azure-core-tracing-opentelemetry",
    "azure-ai-contentsafety",
    "azure-ai-evaluation",
    "opentelemetry-sdk",
    "httpx",
    "openai>=2.20.0",
    "python-dotenv",
    "fastapi",
    "uvicorn",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "E402"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
